requirements_base.txt
otree-core>=1.1.2
Django==1.8.8 # for heroku, needs to be explicitly in requirements file

# Main Django stuffs
django-environ==0.4.1
django-extensions==1.7.5
django-annoying==0.10.3

# Forms
django-braces==1.10.0
django-crispy-forms==1.6.1

# Models
django-model-utils==2.6
django-reversion==2.0.8

# Users
django-allauth==0.31.0

# General utilities
arrow==0.10.0
channels-api==0.3.1


# Make sure debug toolbar gets a version of sqlparse it can use
sqlparse==0.2.2
django-debug-toolbar==1.6
django-channels-panel==0.0.4